requirements.txt
apsw==3.8.6.post1
BBQSQL==1.0
BeautifulSoup==3.2.1
beautifulsoup4==4.4.1
binwalk==2.0.0
BlindElephant==1.0
blinker==1.3
capstone==3.0
certifi==2015.4.28
cffi==0.8.6
characteristic==0.1.0
chardet==2.3.0
Cheetah==2.4.4
CherryPy==3.5.0
chirp==0.4.0
colorama==0.3.2
ConfigArgParse==0.9.3
configobj==5.0.6
configparser==3.3.0.post2
cryptography==0.9.3
cssselect==0.9.1
cssutils==0.9.10
ctypeslib==0.5.6
decorator==3.4.0
defusedxml==0.4.1
dicttoxml==1.6.6
dissy==9
distorm3==3
Django==1.9.1
django-bootstrap-form==3.2
django-bootstrap-pagination==1.5.1
django-htmlmin==0.8.0
djangorestframework==3.3.2
dnspython==1.12.0
docutils==0.12
ecdsa==0.11
Elixir==0.7.1
enum34==1.0.3
feedparser==5.1.3
Flask==0.10.1
flickrapi==1.2
funkload==1.16.1
fuse-python==0.2.1
GeoIP==1.3.2
gevent==1.0.1
greenlet==0.4.2
gyp==0.1
hpack==1.1.0
html2text==2015.6.21
html5lib==0.999
httplib2==0.9
idna==2.0
impacket==0.9.13
ipaddress==1.0.14
IPy==0.81
itsdangerous==0.24
Jinja2==2.7.3
jsonpickle==0.8.0
jsonrpclib==0.1.3
keepnote==0.7.8
killerbee==1.0
lxml==3.4.0
M2Crypto==0.21.1
Magic-file-extensions==0.2
Mako==1.0.0
Markdown==2.5.1
MarkupSafe==0.23
matplotlib==1.4.2
mechanize==0.2.5
mercurial==3.1.2
mitmproxy==0.13
mock==1.0.1
MySQL-python==1.2.3
ndg-httpsclient==0.3.3
netaddr==0.7.12
netifaces==0.10.4
netlib==0.13.1
networkx==1.9
NfSpy==1.0
nose==1.3.4
numpy==1.8.2
oauth==1.0.1
openpyxl==1.7.0
PAM==0.4.2
paramiko==1.15.1
passlib==1.6.1
pbr==1.8.1
pcapy==0.10.8
pefile==1.2.9.1
Pillow==2.6.1
ply==3.4
prettytable==0.7.2
psycopg2==2.5.4
pyasn1==0.1.7
pyasn1-modules==0.0.5
PyBluez==0.18
pycparser==2.10
pycrypto==2.6.1
pycryptopp==0.6.0.1206569328141510525648634803928199668821045408958
pycups==1.9.63
pycurl==7.19.5
pydns==2.3.6
Pygments==2.0.1
pygobject==3.14.0
pygraphviz==1.3rc2
pyinotify==0.9.4
pymssql==1.0.2
PyOpenGL==3.0.2
pyOpenSSL==0.15.1
pyparsing==2.0.3
pyperclip==1.5.11
pyqtgraph==0.0.0
pyregfi==99.99.99.277
pyrit==0.4.0
pyrtlsdr==0.1.1
pyscard==1.6.14
pyserial==2.6
pysmbc==1.0.15.3
pysnmp==4.2.5
pysnmp-apps==0.3.2
pysnmp-mibs==0.1.3
PySocks==1.5.0
pysqlite==2.6.3
python-apt==0.9.3.11
python-dateutil==2.2
python-debian==0.1.27
python-debianbts==1.11
pytidylib==0.2.1
pytz==2012rc0
pyusb==1.0.0a3
PyX==0.12.1
pyxdg==0.25
PyYAML==3.11
pyzmq==14.4.0
reportbug==6.6.3
repoze.lru==0.6
requests==2.7.0
rfidiot==1.0
roman==2.0.0
Routes==2.0
scapy==2.2.0
scipy==0.14.0
service-identity==1.0.0
simplejson==3.6.5
six==1.10.0
slowaes==0.1a1
SOAPpy==0.12.22
SQLAlchemy==0.9.8
stevedore==1.10.0
tornado==4.2
tweepy==2.3
Twisted==14.0.2
Twisted-Conch==14.0.2
Twisted-Core==14.0.2
Twisted-Lore==14.0.2
Twisted-Mail==14.0.2
Twisted-Names==14.0.2
Twisted-News==14.0.2
Twisted-Runner==14.0.2
Twisted-Web==14.0.2
Twisted-Words==14.0.2
urllib3==1.10.4
urwid==1.3.0
uTidylib==0.2
vboxapi==1.0
virtualenv==13.1.2
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
volatility==2.4
wapiti==2.3.0
WebOb==1.4
webunit==1.3.10
Werkzeug==0.9.6
wheel==0.24.0
wstools==0.4.3
wxPython==3.0.1.1
wxPython-common==3.0.1.1
XlsxWriter==0.5.2
xsser==1.6
Yapsy==1.10.423
yara-python==3.1
zenmap==7.1
zim==0.62
zope.interface==4.1.1
